Issue #120: rename ddf to ddf_dvf
